Donation
This has been a dream of mine all my life to produce something that the world can enjoy.  I believe this is the dream come true.  As I close in on all my schooling and rounding the bind of my journey.  I have dedicated all of my time to finishing this project as quickly as possible.  Now the pressure is on from all angles money is tight and nerves are flared and I just pray to God that I can finish before a loved one kills me...