changes
changes are something we all go through as we learn and our belief system is affected by the affects of what we learn.  These are the most powerful of learning opportunities.   So we must in these moment re-evaluate everything and fix what is not correct then before its a problem later.  So through out my time here you will see things change as I learn new things and I hope to do my best to make known what is different whether it is in skill, philosophical, or simply stupidity; I want to do my best to keep it direct for my family and the future and I would love for others to do the same.  That said I would also encourage others to try better to simplify explanations rather then there efforts.  You can never put enough work in that does not better you but we don't try hard enough at the things that better others.  Many things I have read about here on github and in the beginning I understood basically nothing of what people were trying to show each other and that's not to say that I understand much of it now.  I think or my hope as that everyone could take a little more time to be more thorough in explaining what there page is about and is required to perform the task that they have committed.  This in turn should cut many problems down in all fields and create more individuals who understand better and create more solutions.  This is where many knowledgeable, lack wisdom.  They are not the same.  Knowledge is only good to the individual that has it if they lack wisdom.  Without the wisdom the knowledge can not be properly shared with others because wisdom is only found with love.  Without love you can not have wisdom and without those your knowledge is for nothing but you.  <cite>When wisdom enters into your heart and knowledge becomes pleasant to your soul:  Discretion will preserve you, and knowledge will keep you...</cite><u>Prov. 2:10-11.</u>  NOTE:  I would encourage people to read exactly what wisdom will protect you from, I think that a lot of you will find this quite interesting...