Sponsor button
I have added the sponsor button we are now live with it if you would like to sponsor us we would appreciate it.  Note:  I am having trouble with the sponsor code provided It doesn't seem to show up correctly on anything and I don't particularly see anything wrong with it.  When I use it at the W3School site is the only time that it appears correctly and it is a very interesting issue...  I am open to solutions while I figure out what the problem is which I'm sure will turn out to be me somehow lol!